+++ Api/Models/Dtos/Restaurant/RestaurantSummaryVM.cs
@@ -101,7 +101,7 @@ namespace Reservant.Api.Models.Dtos.Restaurant
         /// Rating of the restaurant based on the reviews
         /// </summary>
         [Required]
-        public required decimal Rating { get; set; }
+        public required double Rating { get; set; }
 
         /// <summary>
         /// Number of reviews about this restaurant
